refactor: move files into src & assetss
